Use newer ISO from ipxe for pxeboot
Werewulf seems to require bzImage support for compute nodes
pxebooting. However, the default 'network boot' support in
VirtualBox does not seem to support that, nor does it intend
to (https://www.virtualbox.org/ticket/15159). This causes
compute nodes to not come up.

Based on [this
comment](https://www.virtualbox.org/ticket/15159#comment:4)
I was able to get compute nodes to come up by using the upstream
IPXE pxeboot image. This is the 'full' version of what VirtualBox
uses, and supports bzImage.

This PR requires now that the user download and keep an ISO
handy. However, we could also do the download ourselves, and spare
the user this extra step.